be more careful with os.remove if files do not exist
git-svn-id: https://svn.osgeo.org/gdal/trunk@17052 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/ogr/ogr_gml_read.py
@@ -346,9 +346,16 @@ def ogr_gml_9():
 #  Cleanup
 
 def ogr_gml_cleanup():
+    if not gdaltest.have_gml_reader:
+        return 'skip'
+    
     gdaltest.clean_tmp()
-    os.remove( 'data/utf8.gfs' )
-    os.remove( 'data/ticket_2349_test_1.gfs' )
+    try:
+        os.remove( 'data/utf8.gfs' )
+        os.remove( 'data/ticket_2349_test_1.gfs' )
+    except:
+        pass
+    
     return 'success'
 
 gdaltest_list = [
